+++ application/controllers/wagesmanage.php
@@ -32,7 +32,7 @@ class WagesManage extends M_Controller {
 		$dyn_str = '';
 		for ($i=1;$i<=11;$i++){
 			if( $params['ziduanming'.$i] ){
-				if($params['leixing'.$i] == 'float'){
+				if($params['leixing'.$i] == 'decimal'){
 					$len = '20,2';
 				}else{
 					$len = 50;
@@ -90,12 +90,13 @@ class WagesManage extends M_Controller {
 		$qianchaxun = $this->input->post('qianchaxun',TRUE)?1:0;
 		$houchaxun = $this->input->post('houchaxun',TRUE)?1:0;
 		$summary = $this->input->post('summary',TRUE)?1:0;
-		if($leixing == 'float'){
+		if($leixing == 'decimal'){
 			$len = '20,2';
 		}else{
 			$len = 50;
 		}
 		$sql="ALTER TABLE  `ab22_gongzibiao` CHANGE  `".$field_name."`  `".$new_field_name."` ".$leixing."( ".$len." ) NULL DEFAULT NULL COMMENT  '".$beizhu."'";
+		//$sql="ALTER TABLE  `ab22_gongzibiao` MODIFY COLUMN `".$field_name."`  `".$new_field_name."` ".$leixing."( ".$len." ) NULL DEFAULT NULL COMMENT  '".$beizhu."'";
 		$this->home_model->sqlQuery($sql);
 		
 		$data = array(
