+++ application/views/user_information_info.php
@@ -93,7 +93,7 @@
 			<option value="int" <?php if($row['DATA_TYPE']=="int") echo "selected";?>>int</option>
 
 			<option value="varchar" <?php if($row['DATA_TYPE']=="varchar") echo "selected";?>>varchar</option>
-			<option value="float" <?php if($row['DATA_TYPE']=="float") echo "selected";?>>float</option>
+			<option value="decimal" <?php if($row['DATA_TYPE']=="decimal") echo "selected";?>>decimal</option>
 
 			<option value="text" <?php if($row['DATA_TYPE']=="text") echo "selected";?>>text</option>
